+++ CHANGELOG.rst
@@ -1,10 +1,11 @@
 new version (on deck)
 =====================
 - [ENH] Series toset() functionality #570
-- [INF] Make requirements.txt smaller
+- [ENH] Added option to coalesce function to not delete coalesced columns.
 
 v0.18.2
 =======
+- [INF] Make requirements.txt smaller
 - [ENH] Add a reset_index parameter to shuffle
 - [DOC] Added contribution page link to readme
 - [DOC] fix example for ``update_where``, provide a bit more detail, and expand the bad_values example notebook to demonstrate its use by @anzelpwj.
@@ -21,6 +22,7 @@ v0.18.2
 - [ENH] Add chirality to morgan fingerprints in janitor.chemistry submodule by @Clayton-Springer
 - [INF] ``import_message`` suggests python dist. appropriate installs by @hectormz
 - [ENH] Add count_cumulative_unique() method to janitor.functions submodule by @rahosbach
+- [ENH] Add ``update_where()`` method to ``janitor.spark.functions`` submodule by @zjpoh
 
 v0.18.1
 =======